Added class based streams and refactoring of code
