+++ tap_tiktok_ads/discover.py
@@ -1,7 +1,6 @@
 from singer import Schema, CatalogEntry, Catalog
-
-from tap_tiktok_ads.schemas import get_schemas, STREAMS
-
+from tap_tiktok_ads.schemas import get_schemas
+from tap_tiktok_ads.streams import STREAMS
 
 def discover():
     schemas, field_metadata = get_schemas()
@@ -14,8 +13,8 @@ def discover():
                 tap_stream_id=stream_name,
                 stream=stream_name,
                 schema=schema,
-                key_properties=STREAMS[stream_name]['table_key_properties'],
-                replication_key=STREAMS[stream_name]['valid_replication_keys'],
+                key_properties=STREAMS[stream_name].key_properties,
+                replication_key=STREAMS[stream_name].replication_keys,
                 metadata=mdata
             )
         )
