+++ tap_tiktok_ads/schemas.py
@@ -1,44 +1,11 @@
 import json
 import os
 from singer import metadata
+from tap_tiktok_ads.streams import STREAMS
 
 # Reference:
 #   https://github.com/singer-io/getting-started/blob/master/docs/DISCOVERY_MODE.md#Metadata
 
-STREAMS = {
-    'advertisers': {
-        'table_key_properties': ['id', 'create_time'],
-        'valid_replication_keys': ['create_time']
-    },
-    'campaigns': {
-        'table_key_properties': ['advertiser_id', 'campaign_id', 'modify_time'],
-        'valid_replication_keys': ['modify_time']
-    },
-    'adgroups': {
-        'table_key_properties': ['advertiser_id', 'campaign_id', 'adgroup_id', 'modify_time'],
-        'valid_replication_keys': ['modify_time']
-    },
-    'ads': {
-        'table_key_properties': ['advertiser_id', 'campaign_id', 'adgroup_id', 'ad_id', 'modify_time'],
-        'valid_replication_keys': ['modify_time']
-    },
-    'ad_insights': {
-        'table_key_properties': ['advertiser_id', 'ad_id', 'adgroup_id', 'campaign_id', 'stat_time_day'],
-        'valid_replication_keys': ['stat_time_day']
-    },
-    'ad_insights_by_age_and_gender': {
-        'table_key_properties': ['advertiser_id', 'ad_id', 'adgroup_id', 'campaign_id', 'stat_time_day', 'age', 'gender'],
-        'valid_replication_keys': ['stat_time_day']
-    },
-    'ad_insights_by_country': {
-        'table_key_properties': ['advertiser_id', 'ad_id', 'adgroup_id', 'campaign_id', 'stat_time_day', 'country_code'],
-        'valid_replication_keys': ['stat_time_day']
-    },
-    'ad_insights_by_platform': {
-        'table_key_properties': ['advertiser_id', 'ad_id', 'adgroup_id', 'campaign_id', 'stat_time_day', 'platform'],
-        'valid_replication_keys': ['stat_time_day']
-    }
-}
 
 def get_abs_path(path):
     return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
@@ -55,9 +22,9 @@ def get_schemas():
 
         mdata = metadata.get_standard_metadata(
             schema=schema,
-            key_properties=stream_metadata.get('table_key_properties', None),
-            replication_method=stream_metadata.get('replication_method', None),
-            valid_replication_keys=stream_metadata.get('valid_replication_keys', None,)
+            key_properties=stream_metadata.key_properties,
+            replication_method=stream_metadata.replication_method,
+            valid_replication_keys=stream_metadata.replication_keys
         )
         field_metadata[stream_name] = mdata
